+++ agent/pkg/status/controller/managedclusters/managed_cluster_info_syncer_test.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"testing"
 
+	http "github.com/go-kratos/kratos/v2/transport/http"
+	kessel "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
+	"github.com/project-kessel/inventory-client-go/v1beta1"
 	clusterinfov1beta1 "github.com/stolostron/cluster-lifecycle-api/clusterinfo/v1beta1"
 	"github.com/stretchr/testify/assert"
 	"helm.sh/helm/v3/pkg/time"
@@ -13,14 +16,14 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	"github.com/stolostron/multicluster-global-hub/pkg/transport/requester"
+	"github.com/stolostron/multicluster-global-hub/pkg/transport"
 )
 
 func TestManagedClusterInfoCtrlReconcile(t *testing.T) {
 	// Setup scheme and mock requester
 	scheme := runtime.NewScheme()
 	_ = clusterinfov1beta1.AddToScheme(scheme)
-	mockRequester := &requester.MockRequest{}
+	mockRequester := &MockRequest{}
 
 	// Define test cases
 	creatingTime := metav1.Now()
@@ -96,4 +99,36 @@ func TestManagedClusterInfoCtrlReconcile(t *testing.T) {
 			assert.Equal(t, tt.expectedResult, result)
 		})
 	}
+}
+
+type MockRequest struct{}
+
+func (c *MockRequest) RefreshClient(ctx context.Context, restfulConn *transport.RestfulConfig) error {
+	return nil
+}
+
+func (c *MockRequest) GetHttpClient() *v1beta1.InventoryHttpClient {
+	return &v1beta1.InventoryHttpClient{
+		K8sClusterService: &ClusterServiceClient{},
+	}
+}
+
+type ClusterServiceClient struct{}
+
+func (c *ClusterServiceClient) CreateK8SCluster(ctx context.Context, in *kessel.CreateK8SClusterRequest,
+	opts ...http.CallOption,
+) (*kessel.CreateK8SClusterResponse, error) {
+	return nil, nil
+}
+
+func (c *ClusterServiceClient) UpdateK8SCluster(ctx context.Context, in *kessel.UpdateK8SClusterRequest,
+	opts ...http.CallOption,
+) (*kessel.UpdateK8SClusterResponse, error) {
+	return nil, nil
+}
+
+func (c *ClusterServiceClient) DeleteK8SCluster(ctx context.Context, in *kessel.DeleteK8SClusterRequest,
+	opts ...http.CallOption,
+) (*kessel.DeleteK8SClusterResponse, error) {
+	return nil, nil
 }
